+++ src/main/java/lotto/service/domain/money/Money.java
@@ -1,5 +1,5 @@
 package lotto.service.domain.money;
 
 public interface Money {
-    long getBudget();
+    int getBudget();
 }
